[Change, Add] Add and change different files

+++ Beyond_Imagination/Global/Vector/Vector3.cpp
@@ -8,7 +8,7 @@ Vector3::Vector3(float x, float y, float z)
 	this->z = z;
 }
 
-float Vector3::length()
+double Vector3::length()
 {
 	double temp = x * x + y * y + z * z;
 	return sqrt(temp);
